@@ -1,261 +1,366 @@
 """
-Context management for trace propagation and span relationships.
+Context management for Noveum Trace SDK.
+
+This module handles trace context propagation across function calls,
+async operations, and thread boundaries.
 """
 
-import threading
-import uuid
-from contextvars import ContextVar
+import contextvars
+from collections.abc import Generator
+from contextlib import contextmanager
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, Any, Dict, List, Optional
+from typing import Any, Optional
 
-if TYPE_CHECKING:
-    from .span import Span
-
-# Context variables for async support
-_current_span: ContextVar[Optional["Span"]] = ContextVar("current_span", default=None)
-_current_trace: ContextVar[Optional["TraceContext"]] = ContextVar(
-    "current_trace", default=None
-)
-
-# Thread-local storage for sync support
-_thread_local = threading.local()
+from noveum_trace.core.span import Span
+from noveum_trace.core.trace import Trace
 
 
 @dataclass
 class TraceContext:
-    """Represents a trace context with metadata and span hierarchy."""
+    """
+    Container for trace context information.
 
-    trace_id: str = field(default_factory=lambda: str(uuid.uuid4()))
-    baggage: Dict[str, Any] = field(default_factory=dict)
-    metadata: Dict[str, Any] = field(default_factory=dict)
-    spans: Dict[str, "Span"] = field(default_factory=dict)
+    This class holds the current trace and span context that
+    propagates across function calls and async operations.
+    """
 
-    def __post_init__(self) -> None:
-        """Initialize context after creation."""
-        if not hasattr(_thread_local, "contexts"):
-            _thread_local.contexts = {}
-
-    def add_span(self, span: "Span") -> None:
-        """Add a span to this trace context."""
-        self.spans[str(span.span_id)] = span
-
-    def get_span(self, span_id: str) -> Optional["Span"]:
-        """Get a span by ID."""
-        return self.spans.get(span_id)
-
-    def set_baggage(self, key: str, value: Any) -> None:
-        """Set a baggage item that will be propagated to child spans."""
-        self.baggage[key] = value
-
-    def get_baggage(self, key: str, default: Any = None) -> Any:
-        """Get a baggage item."""
-        return self.baggage.get(key, default)
-
-    def set_metadata(self, key: str, value: Any) -> None:
-        """Set metadata for this trace context."""
-        self.metadata[key] = value
-
-    def get_metadata(self, key: str, default: Any = None) -> Any:
-        """Get metadata from this trace context."""
-        return self.metadata.get(key, default)
-
-    def to_dict(self) -> Dict[str, Any]:
-        """Convert context to dictionary representation."""
-        return {
-            "trace_id": self.trace_id,
-            "baggage": self.baggage,
-            "metadata": self.metadata,
-        }
-
-    @classmethod
-    def from_dict(cls, data: Dict[str, Any]) -> "TraceContext":
-        """Create context from dictionary representation."""
-        return cls(
-            trace_id=data.get("trace_id", str(uuid.uuid4())),
-            baggage=data.get("baggage", {}),
-            metadata=data.get("metadata", {}),
-        )
+    trace: Optional[Trace] = None
+    span: Optional[Span] = None
+    attributes: dict[str, Any] = field(default_factory=dict)
 
 
-class ContextManager:
-    """Manages trace context and span relationships."""
-
-    def __init__(self) -> None:
-        """Initialize the context manager."""
-        self._span_stack: List[Span] = []
-
-    def get_current_span(self) -> Optional["Span"]:
-        """Get the currently active span."""
-        # Try async context first
-        try:
-            span = _current_span.get()
-            if span is not None:
-                return span
-        except LookupError:
-            pass
-
-        # Fall back to thread-local storage
-        if hasattr(_thread_local, "current_span"):
-            return _thread_local.current_span
-
-        return None
-
-    def set_current_span(self, span: Optional["Span"]) -> None:
-        """Set the currently active span."""
-        # Set in async context
-        _current_span.set(span)
-
-        # Also set in thread-local for sync compatibility
-        _thread_local.current_span = span
-
-    def get_current_trace(self) -> Optional[TraceContext]:
-        """Get the current trace context."""
-        # Try async context first
-        try:
-            trace = _current_trace.get()
-            if trace is not None:
-                return trace
-        except LookupError:
-            pass
-
-        # Fall back to thread-local storage
-        if hasattr(_thread_local, "current_trace"):
-            return _thread_local.current_trace
-
-        return None
-
-    def set_current_trace(self, trace: Optional[TraceContext]) -> None:
-        """Set the current trace context."""
-        # Set in async context
-        _current_trace.set(trace)
-
-        # Also set in thread-local for sync compatibility
-        _thread_local.current_trace = trace
-
-    def start_trace(self, trace_id: Optional[str] = None) -> TraceContext:
-        """Start a new trace context."""
-        if trace_id is None:
-            trace_id = str(uuid.uuid4())
-
-        trace = TraceContext(trace_id=trace_id)
-        self.set_current_trace(trace)
-        return trace
-
-    def end_trace(self) -> None:
-        """End the current trace context."""
-        self.set_current_trace(None)
-        self.set_current_span(None)
-
-    def push_span(self, span: "Span") -> None:
-        """Push a span onto the context stack."""
-        self._span_stack.append(span)
-        self.set_current_span(span)
-
-    def pop_span(self) -> Optional["Span"]:
-        """Pop the current span from the context stack."""
-        if not self._span_stack:
-            return None
-
-        span = self._span_stack.pop()
-
-        # Set the previous span as current (if any)
-        if self._span_stack:
-            self.set_current_span(self._span_stack[-1])
-        else:
-            self.set_current_span(None)
-
-        return span
-
-    def get_span_stack(self) -> list:
-        """Get a copy of the current span stack."""
-        return self._span_stack.copy()
+# Context variables for trace propagation
+_trace_context: contextvars.ContextVar[Optional[TraceContext]] = contextvars.ContextVar(
+    "noveum_trace_context", default=None
+)
 
 
-# Global context manager instance
-_context_manager = ContextManager()
+def get_current_context() -> TraceContext:
+    """
+    Get the current trace context.
+
+    Returns:
+        Current TraceContext instance
+    """
+    context = _trace_context.get()
+    if context is None:
+        context = TraceContext()
+        _trace_context.set(context)
+    return context
 
 
-def get_current_span() -> Optional["Span"]:
-    """Get the currently active span."""
-    return _context_manager.get_current_span()
+def get_current_trace() -> Optional[Trace]:
+    """
+    Get the current trace from context.
+
+    Returns:
+        Current Trace instance or None if no trace is active
+    """
+    context = get_current_context()
+    return context.trace
 
 
-def set_current_span(span: Optional["Span"]) -> None:
-    """Set the currently active span."""
-    _context_manager.set_current_span(span)
+def get_current_span() -> Optional[Span]:
+    """
+    Get the current span from context.
+
+    Returns:
+        Current Span instance or None if no span is active
+    """
+    context = get_current_context()
+    return context.span
 
 
-def get_current_trace() -> Optional[TraceContext]:
-    """Get the current trace context."""
-    return _context_manager.get_current_trace()
+def set_current_trace(trace: Optional[Trace]) -> None:
+    """
+    Set the current trace in context.
+
+    Args:
+        trace: Trace instance to set as current
+    """
+    context = get_current_context()
+    context.trace = trace
+    _trace_context.set(context)
 
 
-def set_current_trace(trace: Optional[TraceContext]) -> None:
-    """Set the current trace context."""
-    _context_manager.set_current_trace(trace)
+def set_current_span(span: Optional[Span]) -> None:
+    """
+    Set the current span in context.
+
+    Args:
+        span: Span instance to set as current
+    """
+    context = get_current_context()
+    context.span = span
+    _trace_context.set(context)
 
 
-def start_trace(trace_id: Optional[str] = None) -> TraceContext:
-    """Start a new trace context."""
-    return _context_manager.start_trace(trace_id)
+def set_context_attribute(key: str, value: Any) -> None:
+    """
+    Set an attribute in the current context.
+
+    Args:
+        key: Attribute key
+        value: Attribute value
+    """
+    context = get_current_context()
+    context.attributes[key] = value
+    _trace_context.set(context)
 
 
-def end_trace() -> None:
-    """End the current trace context."""
-    _context_manager.end_trace()
+def get_context_attribute(key: str, default: Any = None) -> Any:
+    """
+    Get an attribute from the current context.
+
+    Args:
+        key: Attribute key
+        default: Default value if key not found
+
+    Returns:
+        Attribute value or default
+    """
+    context = get_current_context()
+    return context.attributes.get(key, default)
 
 
-class SpanContext:
-    """Context manager for automatic span lifecycle management."""
+@contextmanager
+def trace_context(
+    trace: Optional[Trace] = None, span: Optional[Span] = None, **attributes: Any
+) -> Generator[TraceContext, None, None]:
+    """
+    Context manager for setting trace context.
 
-    def __init__(self, span: "Span"):
-        """Initialize span context."""
+    Args:
+        trace: Trace to set as current
+        span: Span to set as current
+        **attributes: Additional context attributes
+
+    Yields:
+        TraceContext instance
+
+    Example:
+        >>> with trace_context(user_id="123", session_id="abc"):
+        ...     # Code here runs with the specified context
+        ...     result = some_function()
+    """
+    # Get current context
+    current_context = get_current_context()
+
+    # Create new context with provided values
+    new_context = TraceContext(
+        trace=trace or current_context.trace,
+        span=span or current_context.span,
+        attributes={**current_context.attributes, **attributes},
+    )
+
+    # Set new context
+    token = _trace_context.set(new_context)
+
+    try:
+        yield new_context
+    finally:
+        # Restore previous context
+        _trace_context.reset(token)
+
+
+@contextmanager
+def span_context(span: Span) -> Generator[Span, None, None]:
+    """
+    Context manager for setting a span as current.
+
+    Args:
+        span: Span to set as current
+
+    Yields:
+        The span instance
+
+    Example:
+        >>> span = trace.create_span("operation")
+        >>> with span_context(span):
+        ...     # Code here runs with span as current
+        ...     result = some_function()
+    """
+    # Get current context
+    current_context = get_current_context()
+
+    # Create new context with the span
+    new_context = TraceContext(
+        trace=current_context.trace,
+        span=span,
+        attributes=current_context.attributes.copy(),
+    )
+
+    # Set new context
+    token = _trace_context.set(new_context)
+
+    try:
+        yield span
+    finally:
+        # Restore previous context
+        _trace_context.reset(token)
+
+
+def copy_context() -> TraceContext:
+    """
+    Create a copy of the current trace context.
+
+    Returns:
+        Copy of current TraceContext
+    """
+    current = get_current_context()
+    return TraceContext(
+        trace=current.trace, span=current.span, attributes=current.attributes.copy()
+    )
+
+
+def clear_context() -> None:
+    """Clear the current trace context."""
+    _trace_context.set(TraceContext())
+
+
+def propagate_context_to_thread(context: Optional[TraceContext] = None) -> TraceContext:
+    """
+    Propagate trace context to a new thread.
+
+    Args:
+        context: Context to propagate (defaults to current context)
+
+    Returns:
+        Context to use in the new thread
+
+    Example:
+        >>> import threading
+        >>>
+        >>> def worker():
+        ...     # Restore context in thread
+        ...     context = propagate_context_to_thread()
+        ...     _trace_context.set(context)
+        ...     # Now the thread has access to the trace context
+        ...
+        >>> thread = threading.Thread(target=worker)
+        >>> thread.start()
+    """
+    if context is None:
+        context = get_current_context()
+
+    return TraceContext(
+        trace=context.trace, span=context.span, attributes=context.attributes.copy()
+    )
+
+
+class ContextualSpan:
+    """
+    A span that automatically manages context.
+
+    This class wraps a Span and automatically sets it as the current
+    span when used as a context manager.
+    """
+
+    def __init__(self, span: Span):
+        """
+        Initialize contextual span.
+
+        Args:
+            span: Span to wrap
+        """
         self.span = span
-        self.previous_span: Optional[Span] = None
+        self._previous_span: Optional[Span] = None
 
-    def __enter__(self) -> "Span":
-        """Enter the span context."""
-        self.previous_span = get_current_span()
-        _context_manager.push_span(self.span)
+    def __enter__(self) -> Span:
+        """Context manager entry."""
+        self._previous_span = get_current_span()
+        set_current_span(self.span)
         return self.span
 
     def __exit__(self, exc_type: Any, exc_val: Any, exc_tb: Any) -> None:
-        """Exit the span context."""
-        _context_manager.pop_span()
-
-        # Mark span as error if exception occurred
+        """Context manager exit."""
         if exc_type is not None:
-            self.span.set_status("error")
-            self.span.set_attribute("error.type", exc_type.__name__)
-            self.span.set_attribute("error.message", str(exc_val))
+            self.span.record_exception(exc_val)
+            self.span.set_status(self.span.status.ERROR, str(exc_val))
 
-        # End the span
-        self.span.end()
+        self.span.finish()
+        set_current_span(self._previous_span)
+
+    def __getattr__(self, name: str) -> Any:
+        """Delegate attribute access to the wrapped span."""
+        return getattr(self.span, name)
 
 
-class AsyncSpanContext:
-    """Async context manager for automatic span lifecycle management."""
+class ContextualTrace:
+    """
+    A trace that automatically manages context.
 
-    def __init__(self, span: "Span"):
-        """Initialize async span context."""
-        self.span = span
-        self.previous_span: Optional[Span] = None
+    This class wraps a Trace and automatically sets it as the current
+    trace when used as a context manager.
+    """
 
-    async def __aenter__(self) -> "Span":
-        """Enter the async span context."""
-        self.previous_span = get_current_span()
-        _context_manager.push_span(self.span)
-        return self.span
+    def __init__(self, trace: Trace):
+        """
+        Initialize contextual trace.
 
-    async def __aexit__(self, exc_type: Any, exc_val: Any, exc_tb: Any) -> None:
-        """Exit the async span context."""
-        _context_manager.pop_span()
+        Args:
+            trace: Trace to wrap
+        """
+        self.trace = trace
+        self._previous_trace: Optional[Trace] = None
 
-        # Mark span as error if exception occurred
+    def __enter__(self) -> Trace:
+        """Context manager entry."""
+        self._previous_trace = get_current_trace()
+        set_current_trace(self.trace)
+        return self.trace
+
+    def __exit__(self, exc_type: Any, exc_val: Any, exc_tb: Any) -> None:
+        """Context manager exit."""
         if exc_type is not None:
-            self.span.set_status("error")
-            self.span.set_attribute("error.type", exc_type.__name__)
-            self.span.set_attribute("error.message", str(exc_val))
+            self.trace.set_status(self.trace.status.ERROR, str(exc_val))
 
-        # End the span
-        self.span.end()
+        self.trace.finish()
+        set_current_trace(self._previous_trace)
+
+    def __getattr__(self, name: str) -> Any:
+        """Delegate attribute access to the wrapped trace."""
+        return getattr(self.trace, name)
+
+
+def attach_context_to_span(span: Span) -> None:
+    """
+    Attach current context attributes to a span.
+
+    Args:
+        span: Span to attach context to
+    """
+    context = get_current_context()
+
+    # Add context attributes to span
+    for key, value in context.attributes.items():
+        span.set_attribute(f"context.{key}", value)
+
+
+def inherit_context_attributes(span: Span, parent_span: Optional[Span] = None) -> None:
+    """
+    Inherit context attributes from parent span.
+
+    Args:
+        span: Span to inherit attributes to
+        parent_span: Parent span to inherit from (defaults to current span)
+    """
+    if parent_span is None:
+        parent_span = get_current_span()
+
+    if parent_span is None:
+        return
+
+    # Inherit specific attributes from parent
+    inheritable_attributes = [
+        "user_id",
+        "session_id",
+        "request_id",
+        "agent_id",
+        "workflow_id",
+    ]
+
+    for attr in inheritable_attributes:
+        context_key = f"context.{attr}"
+        if context_key in parent_span.attributes:
+            span.set_attribute(context_key, parent_span.attributes[context_key])
